requirements.txt
dune-client>=0.1.0
python-dotenv
pandas>=2.0.0
requests 